Renamed files from 'super document' to 'super reader'
