+++ super_editor/test/super_reader/reader_test_tools_test.dart
@@ -2,7 +2,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:super_editor/super_reader_test.dart';
 
 import '../test_tools.dart';
-import 'document_test_tools.dart';
+import 'reader_test_tools.dart';
 
 void main() {
   group("SuperEditor test tools", () {
