+++ super_editor/test/super_reader/super_reader_scrolling_test.dart
@@ -5,7 +5,7 @@ import 'package:super_editor/super_reader_test.dart';
 import 'package:super_editor/super_editor.dart';
 
 import '../test_tools.dart';
-import 'document_test_tools.dart';
+import 'reader_test_tools.dart';
 
 void main() {
   group("SuperDocument scrolling", () {
